Fix SVG not displaying in common web-browsers
The SVG images produced by KiCAD cannot be displayed in common
web-browsers. The reason for that is too large coordinates. As @set-soft
pointed out, the "resolution" is configurable and when we use a lower
resolution the images can be shown again.

Since the change makes any conversion functions dependent on the Plotter
configuration, we migrate the relevant conversion functions into the
Plotter class and we remove those functions from global namespace (as
they no longer represent "general truth").